refactor: vote response 리팩토링
