+++ src/main/java/umc/catchy/domain/vote/api/VoteController.java
@@ -13,12 +13,12 @@ import umc.catchy.domain.category.domain.BigCategory;
 import umc.catchy.domain.vote.dto.request.CreateVoteRequest;
 import umc.catchy.domain.vote.dto.request.PlaceVoteRequest;
 import umc.catchy.domain.vote.dto.request.SubmitVoteRequest;
-import umc.catchy.domain.vote.dto.response.CategoryResponse;
-import umc.catchy.domain.vote.dto.response.GroupPlaceResponse;
-import umc.catchy.domain.vote.dto.response.GroupVoteResultResponse;
-import umc.catchy.domain.vote.dto.response.GroupVoteStatusResponse;
-import umc.catchy.domain.vote.dto.response.VoteResponse;
-import umc.catchy.domain.vote.dto.response.VoteResultResponse;
+import umc.catchy.domain.vote.dto.response.category.CategoryResponse;
+import umc.catchy.domain.vote.dto.response.group.GroupPlaceResponse;
+import umc.catchy.domain.vote.dto.response.group.GroupVoteResultResponse;
+import umc.catchy.domain.vote.dto.response.group.GroupVoteStatusResponse;
+import umc.catchy.domain.vote.dto.response.vote.VoteResponse;
+import umc.catchy.domain.vote.dto.response.vote.VoteResultResponse;
 import umc.catchy.domain.vote.service.VoteService;
 import umc.catchy.global.common.response.BaseResponse;
 import umc.catchy.global.common.response.status.SuccessStatus;
