--- src/main/java/umc/catchy/domain/vote/dto/response/CategoryResult.java
@@ -1,16 +1,0 @@
-package umc.catchy.domain.vote.dto.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CategoryResult {
-    private String category;
-    private int count;
-
-    public CategoryResult(String category, int count) {
-        this.category = category;
-        this.count = count;
-    }
-}
