+++ src/main/java/umc/catchy/domain/vote/dto/response/category/CategoryResult.java
@@ -1,0 +1,3 @@
+package umc.catchy.domain.vote.dto.response.category;
+
+public record CategoryResult(String category, int count) {}
