+++ src/main/java/umc/catchy/domain/vote/dto/response/group/GroupPlaceResponse.java
@@ -1,7 +1,8 @@
-package umc.catchy.domain.vote.dto.response;
+package umc.catchy.domain.vote.dto.response.group;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import umc.catchy.domain.vote.dto.response.place.PlaceResponse;
 
 import java.util.List;
