+++ src/main/java/umc/catchy/domain/vote/dto/response/group/GroupVoteResultResponse.java
@@ -1,0 +1,6 @@
+package umc.catchy.domain.vote.dto.response.group;
+
+import umc.catchy.domain.vote.dto.response.category.CategoryResult;
+import java.util.List;
+
+public record GroupVoteResultResponse(String groupLocation, List<CategoryResult> categories) {}
