+++ src/main/java/umc/catchy/domain/vote/dto/response/vote/MemberVoteStatus.java
@@ -1,12 +1,13 @@
-package umc.catchy.domain.vote.dto.response;
+package umc.catchy.domain.vote.dto.response.vote;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class VotedMemberResponse {
+public class MemberVoteStatus {
     private Long memberId;
     private String nickname;
     private String profileImage;
+    private boolean hasVoted;
 }
