+++ src/main/java/umc/catchy/domain/vote/dto/response/vote/VoteResultResponse.java
@@ -1,0 +1,5 @@
+package umc.catchy.domain.vote.dto.response.vote;
+
+import java.util.List;
+
+public record VoteResultResponse(String status, int totalMembers, List<VoteResult> results) {}
