+++ src/main/java/umc/catchy/domain/vote/service/VoteService.java
@@ -27,17 +27,17 @@ import umc.catchy.domain.vote.domain.Vote;
 import umc.catchy.domain.vote.domain.VoteStatus;
 import umc.catchy.domain.vote.dto.request.CreateVoteRequest;
 import umc.catchy.domain.vote.dto.request.PlaceVoteRequest;
-import umc.catchy.domain.vote.dto.response.CategoryDto;
-import umc.catchy.domain.vote.dto.response.CategoryResponse;
-import umc.catchy.domain.vote.dto.response.CategoryResult;
-import umc.catchy.domain.vote.dto.response.GroupPlaceResponse;
-import umc.catchy.domain.vote.dto.response.GroupVoteResultResponse;
-import umc.catchy.domain.vote.dto.response.GroupVoteStatusResponse;
-import umc.catchy.domain.vote.dto.response.MemberVoteStatus;
-import umc.catchy.domain.vote.dto.response.PlaceResponse;
-import umc.catchy.domain.vote.dto.response.VoteResult;
-import umc.catchy.domain.vote.dto.response.VoteResultResponse;
-import umc.catchy.domain.vote.dto.response.VotedMemberResponse;
+import umc.catchy.domain.vote.dto.response.group.GroupVoteResultResponse;
+import umc.catchy.domain.vote.dto.response.category.CategoryDto;
+import umc.catchy.domain.vote.dto.response.category.CategoryResponse;
+import umc.catchy.domain.vote.dto.response.category.CategoryResult;
+import umc.catchy.domain.vote.dto.response.group.GroupPlaceResponse;
+import umc.catchy.domain.vote.dto.response.group.GroupVoteStatusResponse;
+import umc.catchy.domain.vote.dto.response.vote.MemberVoteStatus;
+import umc.catchy.domain.vote.dto.response.place.PlaceResponse;
+import umc.catchy.domain.vote.dto.response.vote.VoteResult;
+import umc.catchy.domain.vote.dto.response.vote.VoteResultResponse;
+import umc.catchy.domain.vote.dto.response.vote.VotedMemberResponse;
 import umc.catchy.global.common.response.status.ErrorStatus;
 import umc.catchy.global.error.exception.GeneralException;
 import umc.catchy.global.util.SecurityUtil;
